:construction: 예외처리 함수들 모듈로 분리
